Moving load_img to io.utils
